Palette de couleur associée à un sous-ensemble de pays, compatible avec les normes graphiques de la revue OFCE

Palette de couleur associée à un sous-ensemble de pays, compatible avec les normes graphiques de la revue OFCE

Usage

ofce_palette_pays(list_iso3 = NULL, HEX = NULL, RVB = NULL)

Arguments

list_iso3

chr, une liste de nom de pays

HEX

boolean (by default) set TRUE to have the color code in HEX

RVB

boolean set TRUE to have the color code in RVB

Value

une palette de couleur (format HEX)