source_data : exécute le code et cache les données

Cette fonction s'utilise presque comme source et permet d'en accélérer l'exécution par le cache des données.

Usage

source_data(
  name,
  relative = getOption("ofce.source_data.relative"),
  cache_rep = getOption("ofce.source_data.cache_rep"),
  hash = getOption("ofce.source_data.hash"),
  lapse = getOption("ofce.source_data.lapse"),
  force_exec = getOption("ofce.source_data.force_exec"),
  prevent_exec = getOption("ofce.source_data.prevent_exec"),
  metadata = getOption("ofce.source_data.metadata")
)

Arguments

name

(character) le chemin vers le code à exécuter (sans extension .r ou .R), ce chemin doit être relatif au projet (voir relative), bien que une recherche sera effectuée

relative

(character) Si "projet" le chemin est supposé relatif au projet, sinon le chemin sera dans le répertoire de travail (attention il peut changer)

cache_rep

(character) Le chemin du dossier dans lequel sont enregistré les caches (défaut _data)

hash

(boléen) Si TRUE (défaut) un changement dans le code déclenche son exécution

lapse

(character) peut être "never" (défaut) "x hours", "x days", "x weeks", "x months", "x quarters", "x years"

force_exec

(boléen) Si TRUE alors le code est exécuté ($FORCE_EXEC par défaut)

prevent_exec

(boléen) Si TRUE alors le code n'est pas exécuté ($PREVENT_EXEC par défaut), ce flag est prioritaire sur les autres, sauf si il n'y a pas de données en cache

metadata

(boléen) Si TRUE (FALSE par défaut) la fonction retourne une liste avec des métadonnées et le champ data qui contient les données elles même

Value

data (list ou ce que le code retourne)

Details

Le code est exécuté (dans un environnement local) et le résultat est mis en cache. Il est important que le code se termine par un return(les_données). Si return() n'est pas présent dans le code, il n'est pas exécuté et un message d'erreur est envoyé ("NULL" est retourné). le code est exécuté avec un contrôle d'erreur, donc si il bloque, "NULL" est renvoyé, mais sans erreur ni arrêt. les appels suivants seront plus rapides et sans erreur.

Une modification du code est détectée et déclenche l'éxécution

Suivant le paramètre lapse on peut déclencher une exécution péridique. Par exemple, pour ne pas rater une MAJ, on peut mettre lapse = "1 day" ou "day" et une fois par jour le code sera exécuté. Cela permet d'éviter une exécution à chaque rendu.

On peut bloquer l'exécution en renseignant la variable d'environnement "PREVENT_EXEC" pas Sys.setenv(PREVENT_EXEC = "TRUE") ou dans .Renviron. Ce blocage est prioritaire sur tous les autres critères (sauf en cas d'absence de cache).

Des métadonnées peuvent être renvoyées (paramètre metadata) avec la date d'exécution ($date), le temps d'exécution ($timing), la taille des données ($size), le chemin de la source ($where), le hash du source ($hash) et bine sûr les données ($data)

Les valeurs par défaut peuvent être modifiées simplement par options(ofce.source_data.hash = FALSE) par exemple et persiste pendant une session. Typiquement cela peut être mis dans rinit.r (et donc être exécuté par ofce::init_qmd())